Fix problem where original author thought `unsigned long' was the best 32-bit type (rather than define his own type based on the type of box being compiled on).
Submitted by:	Mark Abene <phiber@radicalmedia.com>
		(however I applied a slightly different fix)

+++ lib/libskey/skeysubr.c
@@ -1,3 +1,5 @@
+/* $FreeBSD$ */
+
 #include <err.h>
 #include <stdio.h>
 #include <stdlib.h>
@@ -21,7 +23,7 @@ const char *passwd;   /* Password, any length */
 {
 	char *buf;
 	MDX_CTX md;
-	u_long results[4];
+	u_int32_t results[4];
 	unsigned int buflen;
 
 	buflen = strlen(seed) + strlen(passwd);
@@ -51,7 +53,7 @@ f(x)
 char *x;
 {
 	MDX_CTX md;
-	u_long results[4];
+	u_int32_t results[4];
 
 	MDXInit(&md);
 	MDXUpdate(&md,(unsigned char *)x,8);
